% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_fis.R
\docType{class}
\name{AggregFis}
\alias{AggregFis}
\title{Class "AggregFis"}
\description{
The Fis aggregation operator to be used in \link{Fusion}
}
\section{Slots}{

\describe{
\item{\code{fis}}{\link{Fis} object, The Fis to be used in the aggregation operator}

\item{\code{output_index}}{\link{integer} value, The index (1-based index) of the output in the Fis to be used in the aggregation}
}}

\seealso{
{
\link{NewAggregFis}

\href{https://www.geofis.org/en/documentation-en/data-fusion/#linguistic-rules}{Aggregation using linguistic rules}
}
}
