% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_function.R
\docType{class}
\name{AggregFunction}
\alias{AggregFunction}
\title{Class "AggregFunction"}
\description{
The functional aggregation operator to be used in \link{Fusion}
}
\section{Slots}{

\describe{
\item{\code{func}}{Function, The function used for the aggregation}
}}

\seealso{
\link{NewAggregFunction}
}
