% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_wam.R
\docType{class}
\name{AggregWam}
\alias{AggregWam}
\title{Class "AggregWam"}
\description{
The WAM aggregation operator to be used in \link{Fusion}
}
\section{Slots}{

\describe{
\item{\code{weights}}{\link{numeric} vector, The weights of the WAM aggregation operator (the sum of the weights must be equal to 1 without negative values)}
}}

\seealso{
{
\link{NewAggregWam}

\href{https://www.geofis.org/en/documentation-en/data-fusion/#numerical-operators}{Aggregation using numerical operators}
}
}
