% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductivity_data.R
\docType{data}
\name{conductivity_2014}
\alias{conductivity_2014}
\title{Soil conductivity 2014 dataset}
\format{
\link{SpatialPointsDataFrame} object with 353 observations and 1 attribute:
\describe{
\item{\code{conduct}}{\link{numeric} value, The soil conductivity}
}
}
\usage{
data(conductivity_2014)
}
\description{
The soil conductivity of a vine plot in year 2014
}
\keyword{datasets}
