% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conductivity_data.R
\docType{data}
\name{conductivity_border}
\alias{conductivity_border}
\title{Border dataset}
\format{
\link{SpatialPolygonsDataFrame} object with 1 polygon delimiting the border of the vine plot:
\describe{
\item{\code{id}}{\link{integer} value, The id of the polygon}
}
}
\usage{
data(conductivity_border)
}
\description{
The soil conductivity border of a vine plot
}
\keyword{datasets}
