% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_in_zone.R
\docType{methods}
\name{DataInZone}
\alias{DataInZone}
\title{Classify data in management zones}
\usage{
DataInZone(data, maps, use_id = FALSE)
}
\arguments{
\item{data}{\link{SpatialPointsDataFrame} or \link{SpatialMultiPointsDataFrame} object, The input data.}

\item{maps}{\link{SpatialPolygonsDataFrame} object, or a \link{list} of \link{SpatialPolygonsDataFrame}, The map or list of maps to process.}

\item{use_id}{boolean, Use the id attribute of the zone in the map if TRUE, or the order index (1-based indexed) of the zone in the map if FALSE (the default).}
}
\value{
\link{SpatialPointsDataFrame} or \link{SpatialMultiPointsDataFrame} object depending of the type of input data. Return table of membership of each point in the zones for each map, the input data and a column for each map processed (named "map\{number of zones in the map\}") with the class of each point.
}
\description{
Classify data in management zones of maps obtained with the \link{Zoning} algorithms.
}
