% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\docType{methods}
\name{EuclideanDistance}
\alias{EuclideanDistance}
\title{The "Euclidean" distance}
\usage{
EuclideanDistance()
}
\value{
Euclidean distance object
}
\description{
Function to create an "Euclidean" distance\cr
To be used with the \link{Zoning} \code{combine_distance} or \code{attribute_distance} field
}
