% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion_cars_data.R
\docType{data}
\name{fusion_cars}
\alias{fusion_cars}
\title{Fusion Cars dataset}
\format{
\link{data.frame} object with four cars described by four attributes:
\describe{
\item{\code{A}}{\link{numeric} value, the acceleration time (s) from 0 to 100 km/h}
\item{\code{V}}{\link{numeric} value, the volume of the trunk (l)}
\item{\code{S}}{\link{numeric} value, the maximum speed (km/h)}
\item{\code{C}}{\link{numeric} value, the gas consumption (l per 100 km)}
}
}
\usage{
data(fusion_cars)
}
\description{
Illustration dataset for data fusion numerical operators learning
}
\keyword{datasets}
