% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\docType{methods}
\name{FuzzyDistance}
\alias{FuzzyDistance}
\title{The "Fuzzy" distance}
\usage{
FuzzyDistance(fisin)
}
\arguments{
\item{fisin}{\link{FisIn} object, The partition used for the fuzzy distance (must be a standardized fuzzy partition)}
}
\value{
Fuzzy distance object
}
\description{
Function to create a "Fuzzy" distance\cr
The fuzzy distance function is based on a fuzzy partition that allows for integrating expert knowledge into distance calculations\cr
To be used with the \link{Zoning} \code{attribute_distance} field
}
\references{
{
\insertRef{Infsci13}{GeoFIS}

\insertRef{fuzzieee13}{GeoFIS}
}
}
