% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{GeoFIS}
\alias{GeoFIS-package}
\alias{GeoFIS}
\title{GeoFIS package}
\description{
\href{https://www.geofis.org}{GeoFIS} is an open source software that provides methods for processing spatial data for decision making through a user-friendly interface \insertCite{agriculture8060073}{GeoFIS}.\cr
This R package implements two main functionalities: management zone delineation \insertCite{Pedroso10}{GeoFIS} and data aggregation \insertCite{denys20-ascii,gistam20}{GeoFIS}.
All the mentioned publications are available from the \href{https://www.geofis.org}{GeoFIS} web site.
}
\references{
{
\insertRef{agriculture8060073}{GeoFIS}

\insertRef{fuzzieee12geofis}{GeoFIS}

\insertRef{ijcis13}{GeoFIS}

\insertRef{Pedroso10}{GeoFIS}

\insertRef{denys20-ascii}{GeoFIS}

\insertRef{gistam20}{GeoFIS}
}
}
\seealso{
{
\url{https://www.geofis.org}
}
}
\author{
GeoFIS Team \email{contact@geofis.org}
}
\keyword{internal}
