% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning.R
\name{LearnWamWeights}
\alias{LearnWamWeights}
\title{Learn the WAM weights}
\usage{
LearnWamWeights(data, target, digits = 3)
}
\arguments{
\item{data}{\link{data.frame} or \link{numeric} matrix, The input data (all columns must be in range [0, 1])}

\item{target}{\link{numeric} vector, The target data (must be in range [0, 1])}

\item{digits}{\link{integer} value, The number of digits to which weights are to be rounded (default is 3)}
}
\value{
\link{numeric} vector, The WAM weights
}
\description{
Learn the WAM weights using a non-negative least-square optimization method with the constraint that the sum of weights must be equal to 1.\cr
More information is available in the vignette "Data Fusion with GeoFIS", section "Learning illustration".
}
