% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\docType{methods}
\name{MaximumDistance}
\alias{MaximumDistance}
\title{The "Maximum" distance}
\usage{
MaximumDistance()
}
\value{
Maximum distance object
}
\description{
Function to create a "Maximum" distance\cr
To be used with the \link{Zoning} \code{zone_distance} field
}
