% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\docType{methods}
\name{MeanDistance}
\alias{MeanDistance}
\title{The "Mean" distance}
\usage{
MeanDistance()
}
\value{
Mean distance object
}
\description{
Function to create a "Mean" distance\cr
To be used with the \link{Zoning} \code{zone_distance} field
}
