% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\docType{methods}
\name{MinimumDistance}
\alias{MinimumDistance}
\title{The "Minimum" distance}
\usage{
MinimumDistance()
}
\value{
Minimum distance object
}
\description{
Function to create a "Minimum" distance\cr
To be used with the \link{Zoning} \code{zone_distance} field
}
