% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\docType{methods}
\name{MinkowskiDistance}
\alias{MinkowskiDistance}
\title{The "Minkowski" distance}
\usage{
MinkowskiDistance(power = 2)
}
\arguments{
\item{power}{\link{numeric} value, The power of the Minkowski distance\cr
The default value is 2 (equivalent to euclidean distance)}
}
\value{
Minkowski distance object
}
\description{
Function to create a "Minkowski" distance\cr
To be used with the \link{Zoning} \code{combine_distance} field
}
