% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_fis.R
\name{NewAggregFis}
\alias{NewAggregFis}
\title{Create object of class "AggregFis"}
\usage{
NewAggregFis(fis, output_index = 1)
}
\arguments{
\item{fis}{\link{Fis} object, The Fis to be used in the aggregation operator}

\item{output_index}{\link{integer} value, The index (1-based index) of the output in the Fis to be used in the aggregation (the default is 1)}
}
\value{
\link{AggregFis} object
}
\description{
Function to create an aggregation operator of class \link{AggregFis} to be used in \link{Fusion}
}
\seealso{
\href{https://www.geofis.org/en/documentation-en/data-fusion/#linguistic-rules}{Aggregation using linguistic rules}
}
