% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_function.R
\name{NewAggregFunction}
\alias{NewAggregFunction}
\title{Create object of class "AggregFunction"}
\usage{
NewAggregFunction(func)
}
\arguments{
\item{func}{The function to be used for the aggregation}
}
\description{
Function to create an aggregation operator of class \link{AggregFunction} to be used in \link{Fusion}
}
