% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_owa.R
\name{NewAggregOwa}
\alias{NewAggregOwa}
\title{Create object of class "AggregOwa"}
\usage{
NewAggregOwa(weights)
}
\arguments{
\item{weights}{\link{numeric} vector, The weights of the OWA aggregation operator (the sum of the weights must be equal to 1 without negative values)}
}
\description{
Function to create an aggregation operator of class \link{AggregOwa} to be used in \link{Fusion}
}
\seealso{
\href{https://www.geofis.org/en/documentation-en/data-fusion/#numerical-operators}{Aggregation using numerical operators}
}
