% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_wam.R
\name{NewAggregWam}
\alias{NewAggregWam}
\title{Create object of class "AggregWam"}
\usage{
NewAggregWam(weights)
}
\arguments{
\item{weights}{\link{numeric} vector, The weights of the WAM aggregation operator (the sum of the weights must be equal to 1 without negative values)}
}
\description{
Function to create an aggregation operator of class \link{AggregWam} to be used in \link{Fusion}
}
\seealso{
\href{https://www.geofis.org/en/documentation-en/data-fusion/#numerical-operators}{Aggregation using numerical operators}
}
