% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\docType{methods}
\name{NewFusion}
\alias{NewFusion}
\title{Create object of class "Fusion"}
\usage{
NewFusion(...)
}
\arguments{
\item{...}{arguments of \link{Fusion} constructor}
}
\value{
\link{Fusion} object
}
\description{
Function to create object of class \link{Fusion}
}
