% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\docType{methods}
\name{NewFusionAggreg}
\alias{NewFusionAggreg}
\title{Create an aggregation node to be used in data fusion}
\usage{
NewFusionAggreg(name, aggreg, ...)
}
\arguments{
\item{name}{\link{character} vector, The name of the node}

\item{aggreg}{Aggreg object, The aggregation operator to be used to compute the aggregation of satisfaction degrees\cr
must be an \link{AggregWam}, \link{AggregOwa}, \link{AggregFis} or \link{AggregFunction} object}

\item{...}{\link{Node} objects, The nodes to aggregate\cr
can be an input node built with \link{NewFusionInput} or an aggregate node built with \link{NewFusionAggreg} for a hierarchical aggregation structure}
}
\value{
\link{Node} object
}
\description{
Function to create an aggregation node to be used in \link{Fusion}
}
\seealso{
\href{https://www.geofis.org/en/documentation-en/data-fusion/#aggregation}{Aggregation of the degrees}
}
