% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\docType{methods}
\name{NewFusionInput}
\alias{NewFusionInput}
\title{Create an input node to be used in data fusion}
\usage{
NewFusionInput(name, mf, attribute = name)
}
\arguments{
\item{name}{\link{character} vector, The name of the node}

\item{mf}{\link{Mf} object, The membership function to be used to compute the satisfaction degree of the input}

\item{attribute}{\link{character} vector, The attribute name in the source dataset (default is the same as name)}
}
\value{
\link{Node} object
}
\description{
Function to create an input node to be used in \link{Fusion}
}
\seealso{
\href{https://www.geofis.org/en/documentation-en/data-fusion/#satisfaction-degrees}{From raw data to satisfaction degrees}
}
