% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoning.R
\docType{methods}
\name{NewZoning}
\alias{NewZoning}
\title{Create object of class "Zoning"}
\usage{
NewZoning(...)
}
\arguments{
\item{...}{arguments of \link{Zoning} constructor}
}
\value{
\link{Zoning} object
}
\description{
Function to create object of class \link{Zoning}
}
