% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolima_data.R
\docType{data}
\name{tolima}
\alias{tolima}
\title{Tolima dataset}
\format{
\link{data.frame} object with 30 observations and 8 attributes:
\describe{
\item{\code{Cadmium}}{\link{numeric} value, Cadmium in Soil (ppm)}
\item{\code{pH}}{\link{numeric} value, pH Soil (°pH)}
\item{\code{OM}}{\link{numeric} value, Organic Matter (\%)}
\item{\code{P}}{\link{numeric} value, Available Phosphorus (ppm)}
\item{\code{K}}{\link{numeric} value, Exchangeable Potassium (meq/100 g)}
\item{\code{BalanceGap}}{\link{numeric} value, Balance Gap (\%)}
\item{\code{Ngap_N_OpN}}{\link{numeric} value, N Gap (N/Ntarget)}
\item{\code{Base_S}}{\link{numeric} value, Base Saturation (\%)}
}
}
\usage{
data(tolima)
}
\description{
Soil experimental data in three municipalities of Tolima department in Colombia \insertCite{denys20}{GeoFIS}
}
\references{
{
\insertAllCited{}
}
}
\keyword{datasets}
