% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallest_zone.R
\docType{methods}
\name{ZoneArea}
\alias{ZoneArea}
\title{The "Area" smallest zone}
\usage{
ZoneArea(area)
}
\arguments{
\item{area}{\link{numeric} value, The minimum area of the zone to retain the zone in the \link{Zoning} process}
}
\value{
Area Smallest zone object
}
\description{
Function to create an "Area" smallest zone\cr
To be used with the \link{Zoning} \code{smallest_zone} field
}
