% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallest_zone.R
\docType{methods}
\name{ZoneSize}
\alias{ZoneSize}
\title{The "Size" smallest zone}
\usage{
ZoneSize(number_of_points)
}
\arguments{
\item{number_of_points}{\link{integer} value, The minimum number of points in the zone to retain the zone in the \link{Zoning} process}
}
\value{
Size Smallest zone object
}
\description{
Function to create a "Size" smallest zone\cr
To be used with the \link{Zoning} \code{smallest_zone} field
}
